Fix Makefiles across all C++ modules, write missing exercises
The same rule was broken in all 16 Makefiles (compiled .c instead of
.cpp), so -Wall -Wextra -Werror -std=c++98 was silently never applied.
Fixing it surfaced ~10 real bugs: C++11-only member init syntax invalid
in strict C++98 (ClapTrap, phonebook), signed/unsigned comparisons, a
wrong grade-threshold check in Form (would have let a low-grade AForm
execute an action reserved for a much higher grade), heap corruption in
Array::operator= (allocated one element instead of a full array). Also
wrote the two exercises that had no implementation at all: Account and
Base.

+++ cpp_rank4/cpp00/ex00/src/megaphone.cpp
@@ -23,7 +23,7 @@ int main(int argc, char **argv)
 
 static void megaphone(char *str)
 {
-    for (int i = 0; i < strlen(str); i++)
+    for (size_t i = 0; i < strlen(str); i++)
     {
         if (islower(str[i]))
             str[i] -= 32;
